Simplified use of configuration files.
I found myself a little bit confusing to have a default configuration file
named `config.dist.js` and a local configuration file named `config.js`.
When I say confusing I'm not against the naming, but because how we are
trying to read it at `config-webpack-plugin` usage.

Now we have the following configuration:

```js
  new ConfigPlugin(['./config.js', './config.local.js'])
```

Which is clearer than before. In fact, `config.js` will be the _distributed_
configuration file and `config.local.js` will be the local configuration
file per dev and/or env.

**Important**
Run the following command:

 1. `cp config.js config.local.js`
 2. `cp config.dist.js config.js`

Any question, please drop me a message on Slack.